fix: update of products ingredients/categories now replace all ingredients/categories of the product

+++ src/services/Products/DeleteProduct.service.ts
@@ -1,10 +1,17 @@
 import AppDataSource from "../../data-source";
-import { IParamsIdProduct } from "../../interfaces/Products.interface";
+import AppError from "../../errors/AppError";
 import Product from "../../models/Product.model";
+import { IParamsIdProduct } from "../../interfaces/Products.interface";
 
 class DeleteProductService {
   static async execute({ id }: IParamsIdProduct): Promise<void> {
     const productsRepo = AppDataSource.getRepository(Product);
+
+    const product = await productsRepo.findOne({ where: { id } });
+
+    if (!product) {
+      throw new AppError("Product not found", 404);
+    }
 
     await productsRepo.delete(id);
   }
